+++ packages/sourcepoint_unified_cmp_platform_interface/lib/sourcepoint_unified_cmp_platform_interface.dart
@@ -1,2 +1,2 @@
-export 'package:sourcepoint_unified_cmp_platform_interface/src/interface.dart';
-export 'package:sourcepoint_unified_cmp_platform_interface/src/types.dart';
+export 'src/interface.dart';
+export 'src/types.dart';
